Avoid to pack multiple spine version.
